requirements.txt
matplotlib
numpy
pandas
PyYAML
scikit-learn
scipy
tensorboard
torch
torchaudio
torchvision
torchinfo
